refactor(*): 디렉토리 구조 수정, Error 객체 생성하여 throw
- App 컴포넌트를 src/component로 이동
- 에러 throw 시 `thrwo new Error(message)`
