+++ src/index.js
@@ -1,4 +1,4 @@
-import App from './App.js';
+import App from './components/App.js';
 import Store from './store/Store.js';
 
 const store = new Store('vending-machine');
